requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
GeoAlchemy2
